[K8s-Gen] Refactor KubeType for better generality
