+++ src/main/scala/temple/generate/kube/ast/gen/KubeType.scala
@@ -10,7 +10,7 @@ private[kube] object KubeType {
   /** Header represents the header of a Kubernetes yaml file */
   case class Header(apiVersion: String, kind: String, metadata: Metadata)
 
-  case class Body(spec: DeploymentSpec)
+  case class Body(spec: Spec)
 
   /** A Kubernetes yaml file metadata block */
   case class Metadata(name: String, labels: Labels)
@@ -27,28 +27,4 @@ private[kube] object KubeType {
         "kind" -> Option.when(genType == GenType.Deployment)(if (isDb) "db".asJson else "service".asJson),
       )
   }
-
-  case class DeploymentSpec(replicas: Int, selector: Selector, template: Template)
-
-  case class Selector(matchLabels: Labels)
-
-  case class Template(metadata: Metadata, spec: PodSpec)
-
-  case class PodSpec(
-    hostname: String,
-    containers: Seq[Container],
-    imagePullSecrets: Seq[Secret],
-    restartPolicy: String,
-  )
-
-  case class Container(image: String, name: String, ports: Seq[Port])
-
-  case class Port(containerPort: Int)
-
-  case class Secret(name: String) extends JsonEncodable {
-
-    /** Turn a case class into some key-value pairs in preparation for conversion to a JSON object */
-    override def jsonEntryIterator: IterableOnce[(String, Json)] = Seq("name" -> name.asJson)
-  }
-
 }
